horizonclient: Make ledger_request examples fully qualified
