+++ clients/horizonclient/examples_test.go
@@ -6,6 +6,7 @@ import (
 	"time"
 
 	"github.com/stellar/go/clients/horizonclient"
+	hProtocol "github.com/stellar/go/protocols/horizon"
 	"github.com/stellar/go/protocols/horizon/effects"
 )
 
@@ -59,6 +60,40 @@ func ExampleClient_NextEffectsPage() {
 	for recordsFound {
 		// next page
 		nextPage, err := client.NextEffectsPage(page)
+		if err != nil {
+			fmt.Println(err)
+			return
+		}
+
+		page = nextPage
+		if len(nextPage.Embedded.Records) == 0 {
+			recordsFound = false
+		}
+		fmt.Println(nextPage)
+	}
+}
+
+func ExampleClient_NextLedgersPage() {
+	client := horizonclient.DefaultPublicNetClient
+	// all ledgers
+	ledgerRequest := horizonclient.LedgerRequest{Limit: 20}
+	ledgers, err := client.Ledgers(ledgerRequest)
+	if err != nil {
+		fmt.Println(err)
+		return
+	}
+	fmt.Print(ledgers)
+
+	// get next pages.
+	recordsFound := false
+	if len(ledgers.Embedded.Records) > 0 {
+		recordsFound = true
+	}
+	page := ledgers
+	// get the next page of records if recordsFound is true
+	for recordsFound {
+		// next page
+		nextPage, err := client.NextLedgersPage(page)
 		if err != nil {
 			fmt.Println(err)
 			return
@@ -135,6 +170,40 @@ func ExampleClient_PrevEffectsPage() {
 	}
 }
 
+func ExampleClient_PrevLedgersPage() {
+	client := horizonclient.DefaultPublicNetClient
+	// all ledgers
+	ledgerRequest := horizonclient.LedgerRequest{Limit: 20}
+	ledgers, err := client.Ledgers(ledgerRequest)
+	if err != nil {
+		fmt.Println(err)
+		return
+	}
+	fmt.Print(ledgers)
+
+	// get prev pages.
+	recordsFound := false
+	if len(ledgers.Embedded.Records) > 0 {
+		recordsFound = true
+	}
+	page := ledgers
+	// get the prev page of records if recordsFound is true
+	for recordsFound {
+		// prev page
+		prevPage, err := client.PrevLedgersPage(page)
+		if err != nil {
+			fmt.Println(err)
+			return
+		}
+
+		page = prevPage
+		if len(prevPage.Embedded.Records) == 0 {
+			recordsFound = false
+		}
+		fmt.Println(prevPage)
+	}
+}
+
 func ExampleClient_StreamEffects() {
 	client := horizonclient.DefaultTestNetClient
 	// all effects
@@ -151,6 +220,27 @@ func ExampleClient_StreamEffects() {
 		fmt.Println(e)
 	}
 	err := client.StreamEffects(ctx, effectRequest, printHandler)
+	if err != nil {
+		fmt.Println(err)
+	}
+}
+
+func ExampleClient_StreamLedgers() {
+	client := horizonclient.DefaultTestNetClient
+	// all ledgers from now
+	ledgerRequest := horizonclient.LedgerRequest{}
+
+	ctx, cancel := context.WithCancel(context.Background())
+	go func() {
+		// Stop streaming after 60 seconds.
+		time.Sleep(60 * time.Second)
+		cancel()
+	}()
+
+	printHandler := func(ledger hProtocol.Ledger) {
+		fmt.Println(ledger)
+	}
+	err := client.StreamLedgers(ctx, ledgerRequest, printHandler)
 	if err != nil {
 		fmt.Println(err)
 	}
